move shopify code to lambda-layers
